Fixed bugged html and controllers

+++ src/main/java/com/nnk/springboot/domain/CurvePoint.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Entity
@@ -22,11 +23,11 @@ public class CurvePoint {
 
     @NotNull(message = "Term is mandatory")
     @Digits(integer = 10, fraction = 2, message = "Term must be a valid number")
-    private Double term;
+    private BigDecimal term;
 
     @NotNull(message = "Value is mandatory")
     @Digits(integer = 10, fraction = 2, message = "Value must be a valid number")
-    private Double value;
+    private BigDecimal value;
 
     private Timestamp creationDate;
 }
